NAS-107393 Move config options to dropdown, many other details
