+++ src/app/pages/system/general-settings/general-settings.component.html
@@ -1,8 +1,8 @@
 <div class="wrapper">
-  <div class="title"><h2>General Settings</h2></div>
+  <div class="title"><h2>{{ 'General Settings' | translate }}</h2></div>
     <mat-card class="system-card">
         <div class="title-bar">
-            <h3>{{ supportTitle }}</h3>
+            <h3>{{ supportTitle | translate }}</h3>
         </div>
         <mat-divider></mat-divider>
         <div class="spacer"></div>
@@ -11,10 +11,26 @@
 
     <mat-card class="data-card{{i+1}}" *ngFor="let card of dataCards; index as i; ">
         <div class="title-bar">
-            <h3>{{ card.title }}</h3>
-            <button mat-button color="default" (click)="doAdd(card.id)" class="action-icon">Settings</button>
-        <div class="spacer-sm"></div>
+            <h3>{{ card.title | translate }}</h3>
+            <div class="action-icon">
+              <button mat-button color="default" (click)="doAdd(card.id)">{{ 'Settings' | translate }}</button>
+              <ng-container *ngIf="card.actions">
+                <button name="Actions" mat-icon-button [matMenuTriggerFor]="actionsMenu" class="topbar-button-right"
+                ix-auto ix-auto-type="button" ix-auto-identifier="actions">
+                  <mat-icon>more_vert</mat-icon>
+                </button>
+                <mat-menu #actionsMenu="matMenu">
+                  <button name="card.actions[i].value" mat-menu-item *ngFor="let action of card.actions; index as i"
+                    ix-auto ix-auto-type="option" ix-auto-identifier="card.actions[i].value"
+                    (click)="doCardActions(card.actions[i].value)">
+                    <mat-icon>{{ card.actions[i].icon }}</mat-icon>
+                    <span>{{ card.actions[i].label | translate}}</span>
+                  </button>
+                </mat-menu>
+              </ng-container>
+            </div>
         </div>
+        <div class="spacer-sm"></div>
         <mat-divider></mat-divider>
         <div class="spacer-sm"></div>
         <div *ngFor="let item of card.items" class="card-list-item">
@@ -26,41 +42,41 @@
     <mat-card class="ntp-card">
         <div class="title-bar">
             <h3>{{ ntpTitle }}</h3>
-            <button mat-button color="default" class="action-icon" (click)="doAdd('ntp')">Add</button>
+            <button mat-button color="default" class="action-icon" (click)="doAdd('ntp')">{{ 'Add' | translate }}</button>
         </div>
         <div class="spacer-sm"></div>
         <mat-divider></mat-divider>
         <table *ngIf="dataSource && displayedColumns" mat-table [dataSource]="dataSource" class="ntp-table">
             <ng-container matColumnDef="address">
-                <th mat-header-cell *matHeaderCellDef> Address </th>
+                <th mat-header-cell *matHeaderCellDef> {{ 'Address' | translate}} </th>
                 <td mat-cell *matCellDef="let server" (click)="doAdd('ntp', server.id)"> {{server.address}} </td>
-              </ng-container>
-              <ng-container matColumnDef="burst">
-                <th mat-header-cell *matHeaderCellDef> Burst </th>
-                <td mat-cell *matCellDef="let server" (click)="doAdd('ntp', server.id)"> {{server.burst}} </td>
-              </ng-container>
-              <ng-container matColumnDef="iburst">
-                <th mat-header-cell *matHeaderCellDef> iBurst </th>
-                <td mat-cell *matCellDef="let server" (click)="doAdd('ntp', server.id)"> {{server.iburst}} </td>
-              </ng-container>
-              <ng-container matColumnDef="prefer">
-                <th mat-header-cell *matHeaderCellDef> Prefer </th>
-                <td mat-cell *matCellDef="let server" (click)="doAdd('ntp', server.id)"> {{server.prefer}} </td>
-              </ng-container>
-              <ng-container matColumnDef="minpoll">
-                <th mat-header-cell *matHeaderCellDef> Min. Poll </th>
-                <td mat-cell *matCellDef="let server" (click)="doAdd('ntp', server.id)"> {{server.minpoll}} </td>
-              </ng-container>
-              <ng-container matColumnDef="maxpoll">
-                <th mat-header-cell *matHeaderCellDef> Max. Poll </th>
-                <td mat-cell *matCellDef="let server" (click)="doAdd('ntp', server.id)"> {{server.maxpoll}} </td>
-              </ng-container>
-              <ng-container matColumnDef="actions">
-                <th mat-header-cell *matHeaderCellDef></th>
-                <td mat-cell *matCellDef="let server" (click)="doNTPDelete(server)"><mat-icon>delete</mat-icon></td>
-              </ng-container>
+            </ng-container>
+            <ng-container matColumnDef="burst">
+              <th mat-header-cell *matHeaderCellDef> {{ 'Burst' | translate }} </th>
+              <td mat-cell *matCellDef="let server" (click)="doAdd('ntp', server.id)"> {{server.burst}} </td>
+            </ng-container>
+            <ng-container matColumnDef="iburst">
+              <th mat-header-cell *matHeaderCellDef> {{ 'iBurst' | translate }} </th>
+              <td mat-cell *matCellDef="let server" (click)="doAdd('ntp', server.id)"> {{server.iburst}} </td>
+            </ng-container>
+            <ng-container matColumnDef="prefer">
+              <th mat-header-cell *matHeaderCellDef> {{ 'Prefer' | translate}} </th>
+              <td mat-cell *matCellDef="let server" (click)="doAdd('ntp', server.id)"> {{server.prefer}} </td>
+            </ng-container>
+            <ng-container matColumnDef="minpoll">
+              <th mat-header-cell *matHeaderCellDef> {{ 'Min. Poll' | translate}} </th>
+              <td mat-cell *matCellDef="let server" (click)="doAdd('ntp', server.id)"> {{server.minpoll}} </td>
+            </ng-container>
+            <ng-container matColumnDef="maxpoll">
+              <th mat-header-cell *matHeaderCellDef> {{ 'Max. Poll' | translate }} </th>
+              <td mat-cell *matCellDef="let server" (click)="doAdd('ntp', server.id)"> {{server.maxpoll}} </td>
+            </ng-container>
+            <ng-container matColumnDef="actions">
+              <th mat-header-cell *matHeaderCellDef></th>
+              <td mat-cell *matCellDef="let server" (click)="doNTPDelete(server)"><mat-icon>delete</mat-icon></td>
+            </ng-container>
             <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-            <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+            <tr mat-row *matRowDef="let row; columns: displayedColumns;" class="table-row"></tr>
           </table>
     </mat-card>
 </div>
