+++ src/app/pages/system/general-settings/support/support.component.html
@@ -1,13 +1,16 @@
 <div class="mat-content" fxLayout="column">
   <div id="top-row" class="container" fxLayout="row wrap">
-    <div id="top-right" [fxFlex]="isProductImageTall ? '40%' : '100%'">
+    <div [fxFlex]="isProductImageTall ? '40%' : '100%'">
       <div class="sys-image">
           <app-sys-image
           [product_image]="product_image"
         ></app-sys-image>
       </div>
     </div>
-    <div id="top-left" [fxFlex]="isProductImageTall ? '60%' : '100%'" *ngIf="systemInfo">
+    <div 
+      *ngIf="systemInfo"
+      [fxFlex]="isProductImageTall ? '60%' : '100%'" 
+      [class]="extraMargin ? 'extra-margin' : 'regular-margin'">
         <app-sys-info
         *ngIf="!hasLicense; else licensedSysInfo"
         [version]='systemInfo.version'
@@ -36,6 +39,9 @@
         <div class="links" *ngFor="let link of links">
           <div [innerHTML]=link></div>
         </div>
+      </div>
+      <div *ngIf="hasLicense">
+        <mat-checkbox (change)="updateProductionStatus($event)">{{ 'This is a production system' | translate}}</mat-checkbox>
       </div>
     </div>
   </div>
