+++ src/app/services/docs.service.ts
@@ -17,19 +17,22 @@ export class DocsService {
                 message = message.replace(replace, urls[url]);
             }
             const running_version = window.localStorage.getItem('running_version');
-            const web_version = "?runningversion=" + running_version;
-            const version = running_version.split('-');
-            version.shift();
-            const doc_version = version.join('-');
-            if (version && version.length > 1) {
-                message = message.replace(/--version--/g, doc_version);
-            }
-            if (web_version) {
-                message = message.replace(/--webversion--/g, web_version);
-            }
             if (running_version) {
-                message = message.replace(/--runningversion--/g, running_version);
+                const web_version = "?runningversion=" + running_version;
+                const version = running_version.split('-');
+                version.shift();
+                const doc_version = version.join('-');
+                if (version && version.length > 1) {
+                    message = message.replace(/--version--/g, doc_version);
+                }
+                if (web_version) {
+                    message = message.replace(/--webversion--/g, web_version);
+                }
+                if (running_version) {
+                    message = message.replace(/--runningversion--/g, running_version);
+                }
             }
+
 
             const product_type = window.localStorage.getItem('product_type');
             message = message.replace(/--nas--/g, `truenas ${product_type}`);
